refactor: step3 피드백 반영

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,16 +1,14 @@
 package nextstep.subway.favorite.application;
 
 import nextstep.subway.exception.NoSearchFavoriteException;
-import nextstep.subway.exception.NoSearchMemberException;
-import nextstep.subway.exception.NoSearchStationException;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
+import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.domain.Member;
-import nextstep.subway.member.domain.MemberRepository;
+import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.domain.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,31 +19,23 @@ import java.util.stream.Collectors;
 @Transactional(readOnly = true)
 public class FavoriteService {
     private final FavoriteRepository favoriteRepository;
-    private final MemberRepository memberRepository;
-    private final StationRepository stationRepository;
+    private final MemberService memberService;
+    private final StationService stationService;
 
-    public FavoriteService(FavoriteRepository favoriteRepository, MemberRepository memberRepository,
-                           StationRepository stationRepository) {
+    public FavoriteService(FavoriteRepository favoriteRepository, MemberService memberService,
+                           StationService stationService) {
         this.favoriteRepository = favoriteRepository;
-        this.memberRepository = memberRepository;
-        this.stationRepository = stationRepository;
+        this.memberService = memberService;
+        this.stationService = stationService;
     }
 
     @Transactional
     public FavoriteResponse createFavorite(Long loginMemberId, FavoriteRequest favoriteRequest) {
-        Member persistMember = findMemberById(loginMemberId);
-        Station source = findStationById(favoriteRequest.getSourceId());
-        Station target = findStationById(favoriteRequest.getTargetId());
+        Member persistMember = memberService.findMemberById(loginMemberId);
+        Station source = stationService.findStationById(favoriteRequest.getSourceId());
+        Station target = stationService.findStationById(favoriteRequest.getTargetId());
         Favorite persistFavorite = favoriteRepository.save(Favorite.of(persistMember, source, target));
         return FavoriteResponse.of(persistFavorite);
-    }
-
-    private Member findMemberById(Long loginMemberId) {
-        return memberRepository.findById(loginMemberId).orElseThrow(() -> new NoSearchMemberException(loginMemberId));
-    }
-
-    private Station findStationById(Long stationId) {
-        return stationRepository.findById(stationId).orElseThrow(() -> new NoSearchStationException(stationId));
     }
 
     public List<FavoriteResponse> getAllFavorites(Long loginMemberId) {
@@ -58,7 +48,7 @@ public class FavoriteService {
     @Transactional
     public void deleteFavorite(Long loginMemberId, Long favoriteId) {
         Favorite favorite = findFavoriteById(favoriteId);
-        favorite.vlidateOwner(loginMemberId);
+        favorite.validateOwner(loginMemberId);
         favoriteRepository.deleteById(favoriteId);
     }
